Map existing object with source.
Added an extra Mapper.AddMap() method so that we can register a mapping
to map an existing target with the source object and an optional tag.
Added an extra Mapper.Map() method so that it is possible to map an
existing target with a source object and an optional tag.